tmep

+++ modules/lmscan-agent/src/main/scala/io/leisuremeta/chain/lmscan/agent/entity/AccountEntity.scala
@@ -6,8 +6,8 @@ import java.time.Instant
 
 final case class AccountEntity(
     address: String,
-    balance: Long,
-    amount: Long,
+    balance: BigDecimal,
+    amount: BigDecimal,
     eventTime: Long,
     createdAt: Long,
 )
